+++ opsml/registry/cards/model.py
@@ -12,7 +12,6 @@ from numpy.typing import NDArray
 from pydantic import ConfigDict, model_validator
 
 from opsml.helpers.logging import ArtifactLogger
-from opsml.model.model_types import ModelType
 from opsml.model.predictor import OnnxModelPredictor
 from opsml.model.types import (
     ApiDataSchemas,
@@ -21,13 +20,12 @@ from opsml.model.types import (
     ModelMetadata,
     ModelReturn,
     OnnxModelDefinition,
-    TrainedModelType,
     ValidModelInput,
 )
 from opsml.registry.cards.audit_deco import auditable
 from opsml.registry.cards.base import ArtifactCard
 from opsml.registry.cards.types import CardType, ModelCardMetadata
-from opsml.registry.data.types import AllowedDataType, check_data_type
+from opsml.registry.cards.validator import ModelCardValidator
 from opsml.registry.sql.records import ModelRegistryRecord, RegistryRecord
 from opsml.registry.storage.artifact_storage import load_record_artifact_from_storage
 from opsml.registry.storage.types import ArtifactStorageSpecs, ArtifactStorageType
@@ -35,98 +33,6 @@ from opsml.registry.utils.settings import settings
 
 logger = ArtifactLogger.get_logger()
 storage_client = settings.storage_client
-
-
-class ModelCardValidator:
-    def __init__(
-        self,
-        sample_data: ValidModelInput,
-        trained_model: Any,
-        metadata: Optional[ModelCardMetadata] = None,
-    ) -> None:
-        self.sample_data = sample_data
-        self.trained_model = trained_model
-        self.metadata = metadata
-        self.model_class = self._get_model_class_name()
-
-    def _get_model_class_name(self) -> str:
-        """Gets class name from model"""
-
-        if "keras.engine" in str(self.trained_model):
-            return TrainedModelType.TF_KERAS.value
-
-        if "torch" in str(self.trained_model.__class__.__bases__):
-            return TrainedModelType.PYTORCH.value
-
-        # for transformer models from huggingface
-        if "transformers.models" in str(self.trained_model.__class__.__bases__):
-            return TrainedModelType.TRANSFORMER.value
-
-        return str(self.trained_model.__class__.__name__)
-
-    def get_sample(self) -> Optional[Union[pd.DataFrame, NDArray[Any], Dict[str, NDArray[Any]]]]:
-        """Check sample data and returns one record to be used
-        during ONNX conversion and validation
-
-        Returns:
-            Sample data with only one record
-        """
-        if self.sample_data is None:
-            return self.sample_data
-
-        if not isinstance(self.sample_data, dict):
-            if isinstance(self.sample_data, pl.DataFrame):
-                self.sample_data = self.sample_data.to_pandas()
-
-            return self.sample_data[0:1]
-
-        sample_dict = {}
-        if isinstance(self.sample_data, dict):
-            for key in self.sample_data.keys():
-                sample_dict[key] = self.sample_data[key][0:1]
-
-            return sample_dict
-
-        raise ValueError("Provided sample data is not a valid type")
-
-    def get_model_type(self, model_class: str) -> str:
-        model_type = next(
-            (
-                model_type
-                for model_type in ModelType.__subclasses__()
-                if model_type.validate(model_class_name=model_class)
-            )
-        )
-        return model_type.get_type()
-
-    def get_metadata(self) -> ModelCardMetadata:
-        """Checks metadata for valid values
-        Returns:
-            `ModelCardMetadata` with updated sample_data_type
-        """
-
-        model_class = self._get_model_class_name()
-        model_type = self.get_model_type(model_class)
-        data_type = check_data_type(self.sample_data)
-
-        if self.metadata is None:
-            if data_type in [AllowedDataType.IMAGE]:
-                raise ValueError(
-                    f"""Invalid model data input type. Accepted types are a pandas dataframe, 
-                                 numpy array and dictionary of numpy arrays. Received {data_type}""",
-                )
-            self.metadata = ModelCardMetadata(
-                sample_data_type=data_type,
-                model_class=model_class,
-                model_type=model_type,
-            )
-
-        elif self.metadata is not None:
-            self.metadata.sample_data_type = data_type
-            self.metadata.model_type = model_type
-            self.metadata.model_class = model_class
-
-        return self.metadata
 
 
 @auditable
